+++ app/dev/entry.js
@@ -1,2 +1,6 @@
-require('./app.js')
-require ('./controllers.js')();
+require('./app.js');
+require('./controllers/appController.js');
+require('./controllers/attributesController.js');
+require('./controllers/componentsController.js')
+require('./controllers/capabilitiesController.js');
+require('./controllers/projectsListController.js');
