--- app/dev/projectsListContoller.js
@@ -1,1 +1,0 @@
-
